requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch
torchvision
numpy
opencv-python
ultralytics
jkp-depth-anything-v2